Initial Push to private Git Hub Repo
Lantern will be a first-person puzzle game that takes place in a dark dungeon and the goal is to escape by solving puzzles using the provided lantern in the starting room. Some of the features will include throwing the lantern to reveal aspects from a distance and changing the color of the light to reveal other aspects of the world that can’t be seen with the normal light. Overall, the goal for this game is to create something that newer players can easily understand the mechanics, experienced players would have a second option in what they play and a game that is inviting for many to try.