Code Drp 2
This code release continues to focus on the gameplay aspect of Lantern. The new feature that has been implemented into this project is the ability to pick-up the Lantern by clicking the right mouse button, when prompted. The next feature that has also been implemented into this project is that when the Lantern has been picked-up, by using the scroll wheel up or down will change the color of the lantern in a loop starting from yellow moving to blue than green. Lastly, the main light that emanates from the lantern is centered and follows the camera’s position.